CVS: ---------------- CTMD Task 7 Commit (Titanium.imts.us) ------------- CVS: Please format log message lines to be no longer than the above line. Fixing Compile error in the IAttribute.h and subclasses. Return value was different between the superclass and child class.

+++ src/Base/IAttribute.h
@@ -25,7 +25,7 @@ class IAttributeWriter;
  * templated.
  * @author Mike Jackson
  * @date March 2007
- * @version $Revision: 1.5 $
+ * @version $Revision: 1.6 $
  *  
  */
 class MXA_EXPORT IAttribute 
@@ -52,12 +52,12 @@ class MXA_EXPORT IAttribute
     /**
      * @brief Converts the data value to a float value
      */
-    virtual float valueToFloat32() = 0;
+    virtual float32 valueToFloat32() = 0;
 	
     /**
      * @brief Converts the data value to a double value
      */
-    virtual float valueToFloat64() = 0;
+    virtual float64 valueToFloat64() = 0;
     
     /**
      * @brief Converts the data value to a string
